tests: 123-lua-path.t: fixed a few test cases failing in EC2 to-ngx_stream_lua suite.
In to-ngx_lua, the paths to lua-resty-core and lua-resty-lrucache are
not included by default, but is still required for Nginx to start due to
the default value of `lua_load_resty_core`.

Hence, we always provide those paths via the `LUA_PATH` env variable. In
non-to-* suites, those paths will thus be present twice, so the regexes
take that into consideration.